Add files via upload
Thought it might be helpful to do some exploring in the amazon.csv file.
The largely positive reviews based on ratings which seem a bit suspicious, but I don't think that will be a big deal for the project. Might actually interesting to compare our findings to this more simple method.